Extract sync delta merge shape helpers
