+++ tests/test-sync.js
@@ -45,6 +45,7 @@ await import('../js/settings.js');
   const syncDeltaPlannersSrc = await fetchWithRetry('js/sync-delta-planners.js');
   const syncDeltaSnapshotSrc = await fetchWithRetry('js/sync-delta-snapshot.js');
   const syncDeltaMergeSrc = await fetchWithRetry('js/sync-delta-merge.js');
+  const syncDeltaMergeShapesSrc = await fetchWithRetry('js/sync-delta-merge-shapes.js');
   const syncDeltaRegistrySrc = await fetchWithRetry('js/sync-delta-registry.js');
   const syncDeltaObservabilitySrc = await fetchWithRetry('js/sync-delta-observability.js');
   const syncTombstonesSrc = await fetchWithRetry('js/sync-tombstones.js');
@@ -75,7 +76,7 @@ await import('../js/settings.js');
   const stylesSrc = await fetchWithRetry('styles.css');
   const themeExtraSrc = await fetchWithRetry('themes-extra.css');
   const serviceWorkerSrc = await fetchWithRetry('service-worker.js');
-  const deltaSearchSrc = `${syncSrc}\n${syncPushSrc}\n${syncPushDeltasSrc}\n${syncReconcileSrc}\n${syncPullSrc}\n${syncPullMergeSrc}\n${syncCutoverSrc}\n${syncDeltaSrc}\n${syncDeltaPlannersSrc}\n${syncDeltaSnapshotSrc}\n${syncDeltaMergeSrc}\n${syncDeltaRegistrySrc}\n${syncDeltaObservabilitySrc}\n${syncDiagnosticsSrc}\n${syncDiagnoseActionsSrc}\n${syncDiagnoseUiSrc}\n${syncWindowBindingsSrc}`;
+  const deltaSearchSrc = `${syncSrc}\n${syncPushSrc}\n${syncPushDeltasSrc}\n${syncReconcileSrc}\n${syncPullSrc}\n${syncPullMergeSrc}\n${syncCutoverSrc}\n${syncDeltaSrc}\n${syncDeltaPlannersSrc}\n${syncDeltaSnapshotSrc}\n${syncDeltaMergeSrc}\n${syncDeltaMergeShapesSrc}\n${syncDeltaRegistrySrc}\n${syncDeltaObservabilitySrc}\n${syncDiagnosticsSrc}\n${syncDiagnoseActionsSrc}\n${syncDiagnoseUiSrc}\n${syncWindowBindingsSrc}`;
   const exportBlockIncludes = (src, names) => [...src.matchAll(/export\s+\{([^}]*)\};/g)]
     .some(([, block]) => names.every(name => new RegExp(`\\b${name}\\b`).test(block)));
 
@@ -209,10 +210,17 @@ await import('../js/settings.js');
   assert('sync-delta-merge.js owns pull-side itemRow merge overlay',
     syncDeltaMergeSrc.includes('export function configureSyncDeltaMerge')
       && syncDeltaMergeSrc.includes('export async function _mergeItemRowsIntoImported')
-      && syncDeltaMergeSrc.includes('recordPullDeltaSurface')
+      && syncDeltaMergeSrc.includes("from './sync-delta-merge-shapes.js'")
       && syncDeltaMergeSrc.includes('resetPullDeltaSnapshot'));
   assert('service worker precaches sync-delta-merge.js',
     serviceWorkerSrc.includes("'/js/sync-delta-merge.js'"));
+  assert('sync-delta-merge-shapes.js owns pull-side scalar/map/array merge branches',
+    syncDeltaMergeShapesSrc.includes('export async function mergeScalarRowsIntoImported')
+      && syncDeltaMergeShapesSrc.includes('export async function mergeMapRowsIntoImported')
+      && syncDeltaMergeShapesSrc.includes('export async function mergeArrayRowsIntoImported')
+      && syncDeltaMergeShapesSrc.includes('recordPullDeltaSurface'));
+  assert('service worker precaches sync-delta-merge-shapes.js',
+    serviceWorkerSrc.includes("'/js/sync-delta-merge-shapes.js'"));
   assert('sync-delta-registry.js owns delta surfaces and identity config',
     syncDeltaRegistrySrc.includes('export const DELTA_ARRAYS')
       && syncDeltaRegistrySrc.includes('export const DELTA_MAPS')
@@ -229,7 +237,7 @@ await import('../js/settings.js');
       && syncDeltaObservabilitySrc.includes('export function resetDeltaTelemetry')
       && syncDeltaObservabilitySrc.includes('export function getDeltaCutoverReadiness')
       && syncDeltaMergeSrc.includes('resetPullDeltaSnapshot(profileId)')
-      && syncDeltaMergeSrc.includes('recordPullDeltaSurface(arrayName'));
+      && syncDeltaMergeShapesSrc.includes('recordPullDeltaSurface(arrayName'));
   assert('service worker precaches sync-delta-observability.js',
     serviceWorkerSrc.includes("'/js/sync-delta-observability.js'"));
   assert('sync-tombstones.js owns remote profile delete helpers',
@@ -670,20 +678,20 @@ await import('../js/settings.js');
   assert('onSyncReceived overlays per-row state AFTER blob merge',
     /merged\s*=\s*localImportedForMerge[\s\S]{0,400}mergeImportedData[\s\S]{0,800}_mergeItemRowsIntoImported/.test(syncPullMergeSrc));
   assert('_mergeItemRowsIntoImported drops tombstoned items from imported arrays',
-    /_mergeItemRowsIntoImported[\s\S]{0,15000}let nextArr\s*=\s*curArr\.filter\(it\s*=>\s*!tombs\.has\(itemIdFn\(it\)\)\)/.test(syncDeltaMergeSrc));
+    /mergeArrayRowsIntoImported[\s\S]{0,12000}let nextArr\s*=\s*curArr\.filter\(it\s*=>\s*!tombs\.has\(itemIdFn\(it\)\)\)/.test(syncDeltaMergeShapesSrc));
   // Resurrection-prevention seed: blob-side `_deleted[arrayName]` must
   // pre-populate the row-side tombs Set, otherwise a peer pushing the
   // row back as live (before pulling our delete) re-inserts it locally.
   assert('_mergeItemRowsIntoImported seeds tombs from local blob _deleted before walking rows',
-    /imported\.\s*_deleted[\s\S]{0,200}\[arrayName\][\s\S]{0,200}tombs\.add/.test(syncDeltaMergeSrc));
+    /imported\.\s*_deleted[\s\S]{0,200}\[arrayName\][\s\S]{0,200}tombs\.add/.test(syncDeltaMergeShapesSrc));
   assert('_mergeItemRowsIntoImported skips inserting items that match a blob-tombstoned itemId',
-    /tombs\.has\(itemId\)\)\s*continue/.test(syncDeltaMergeSrc));
+    /tombs\.has\(itemId\)\)\s*continue/.test(syncDeltaMergeShapesSrc));
   assert('_mergeItemRowsIntoImported prefers per-row payload when itemId already present in array (replace)',
-    /idx\s*!==\s*undefined[\s\S]{0,200}nextArr\[idx\]\s*=\s*item/.test(syncDeltaMergeSrc));
+    /idx\s*!==\s*undefined[\s\S]{0,200}nextArr\[idx\]\s*=\s*item/.test(syncDeltaMergeShapesSrc));
   assert('_mergeItemRowsIntoImported gunzips GZ|v1| payloads via capped variant',
-    /json\.startsWith\('GZ\|v1\|'\)[\s\S]{0,300}_gunzipToStringCapped\(_base64ToBytes\(json\.slice\(6\)\)\)/.test(syncDeltaMergeSrc));
+    /json\.startsWith\('GZ\|v1\|'\)[\s\S]{0,300}_gunzipToStringCapped\(_base64ToBytes\(json\.slice\(6\)\)\)/.test(syncDeltaMergeShapesSrc));
   assert('_mergeItemRowsIntoImported guards against itemId/payload mismatch (defence-in-depth)',
-    /itemIdFn\(item\)\s*===\s*row\.itemId/.test(syncDeltaMergeSrc));
+    /itemIdFn\(item\)\s*===\s*row\.itemId/.test(syncDeltaMergeShapesSrc));
 
   // Snapshot persistence contract
   assert('Delta snapshot key namespaced per (profile, arrayName)',
@@ -1102,9 +1110,9 @@ await import('../js/settings.js');
   assert('_planArrayDelta uses itemIdFn-derived id everywhere (not item.id)',
     /tuples\s*=\s*Array\.isArray\(items\)[\s\S]{0,300}itemIdFn\(it\)/.test(deltaSearchSrc));
   assert('_mergeItemRowsIntoImported uses itemIdFn for replace-or-insert match',
-    /_mergeItemRowsIntoImported[\s\S]{0,12000}DELTA_ARRAY_CONFIG\[arrayName\][\s\S]{0,25000}itemIdFn\(nextArr\[i\]\)/.test(deltaSearchSrc));
+    /mergeArrayRowsIntoImported[\s\S]{0,8000}DELTA_ARRAY_CONFIG\[arrayName\][\s\S]{0,12000}itemIdFn\(nextArr\[i\]\)/.test(syncDeltaMergeShapesSrc));
   assert('_mergeItemRowsIntoImported verifies payload itemId matches row column',
-    /itemIdFn\(item\)\s*===\s*row\.itemId/.test(syncDeltaMergeSrc));
+    /itemIdFn\(item\)\s*===\s*row\.itemId/.test(syncDeltaMergeShapesSrc));
 
   // Live: round-trip the changeHistory itemIdFn — verify a synth itemId
   // for a realistic recordChange entry is allowlist-safe and stable.
@@ -1389,7 +1397,7 @@ await import('../js/settings.js');
   assert('getDeltaCutoverReadiness walks dotted DELTA_ARRAYS entries via getAt',
     /getDeltaCutoverReadiness[\s\S]{0,2000}arrayName\.includes\('\.'\)[\s\S]{0,200}getAt\(importedData,\s*arrayName\)/.test(deltaSearchSrc));
   assert('_mergeItemRowsIntoImported writes nested arrays back via setAt',
-    /_mergeItemRowsIntoImported[\s\S]{0,12000}isNested\s*=\s*arrayName\.includes\('\.'\)[\s\S]{0,400}setAt\(imported,\s*arrayName,/.test(deltaSearchSrc));
+    /mergeArrayRowsIntoImported[\s\S]{0,1600}isNested\s*=\s*arrayName\.includes\('\.'\)[\s\S]{0,400}setAt\(imported,\s*arrayName,/.test(syncDeltaMergeShapesSrc));
   assert('DELTA_SCALARS includes free-form text fields',
     /'interpretiveLens'/.test(deltaSearchSrc) && /'contextNotes'/.test(deltaSearchSrc));
   assert('_planScalarDelta defined',
@@ -1719,7 +1727,8 @@ await import('../js/settings.js');
   assert('_gunzipToStringCapped throws on cap exceeded',
     /total\s*>\s*maxBytes[\s\S]{0,300}refusing to trust/.test(syncPayloadSrc));
   assert('All per-row gunzip sites use the capped variant',
-    (deltaSearchSrc.match(/_gunzipToStringCapped\(_base64ToBytes\(json\.slice\(6\)\)\)/g) || []).length >= 4);
+    /async function decodeRowPayload[\s\S]{0,500}_gunzipToStringCapped\(_base64ToBytes\(json\.slice\(6\)\)\)/.test(syncDeltaMergeShapesSrc)
+      && (syncDeltaMergeShapesSrc.match(/await decodeRowPayload\(row\)/g) || []).length >= 3);
   // Blob path also routes through _gunzipToStringCapped, with the
   // 5 MB MAX_SYNC_PAYLOAD_BYTES cap — a single capped helper is the
   // only gunzip entry point post-2026-05-10 audit (the bare
